requirements.txt
chainlit==0.7.0
openai==0.28.0
python-dotenv==1.0.0
numpy
matplotlib
plotly
pandas
scipy
scikit-learn
wandb